Move src/helpers to src/globalHelpers
