+++ src/globalHelpers.js
